Separate routing files by path
