+++ src/api/common/schema.ts
@@ -1,0 +1,10 @@
+import { z } from '@hono/zod-openapi';
+
+export const ErrorSchema = z.object({
+  code: z.number().openapi({
+    example: 400,
+  }),
+  message: z.string().openapi({
+    example: 'Bad Request',
+  }),
+});
