+++ src/api/root/route.ts
@@ -1,0 +1,17 @@
+import { createRoute } from '@hono/zod-openapi';
+import { WelcomeSchema } from './schema';
+
+export const welcomeRoute = createRoute({
+  method: 'get',
+  path: '/',
+  responses: {
+    200: {
+      content: {
+        'application/json': {
+          schema: WelcomeSchema,
+        },
+      },
+      description: 'Hello Hono!',
+    },
+  },
+});
