+++ src/api/root/schema.ts
@@ -1,0 +1,7 @@
+import { z } from '@hono/zod-openapi';
+
+export const WelcomeSchema = z.object({
+  message: z.string().openapi({
+    example: 'Hi!',
+  }),
+});
