+++ src/index.ts
@@ -1,10 +1,9 @@
-import { getPrisma } from './prismaFunction';
 import { OpenAPIHono } from '@hono/zod-openapi';
 import { cors } from 'hono/cors';
-import { welcomeRoute, noteRoute, createNoteRoute } from './route';
+import { rootApp } from './api/root';
+import { noteApp } from './api/notes';
 
 type Bindings = {
-  DATABASE_URL: string;
   CORS_ORIGIN: string;
 };
 
@@ -18,65 +17,15 @@ app.use('/*', async (c, next) => {
   })(c, next);
 });
 
-app.openapi(welcomeRoute, (c) => {
-  return c.json({ message: 'Welcome to SprintPost!' }, 200);
-});
-
-app.openapi(
-  noteRoute,
-  async (c) => {
-    const { noteKey } = c.req.valid('param');
-    const prisma = getPrisma(c.env.DATABASE_URL);
-    const note = await prisma.note.findFirst({
-      select: {
-        noteKey: true,
-        subject: true,
-        content: true,
-      },
-      where: {
-        noteKey,
-      },
-    });
-    return note
-      ? c.json(note, 200)
-      : c.json({ code: 404, message: 'Not Found' }, 404);
-  },
-  (result, c) => {
-    if (!result.success) {
-      return c.json({ code: 400, message: 'Validation Error' }, 400);
-    }
-  },
-);
-
-app.openapi(
-  createNoteRoute,
-  async (c) => {
-    const { noteKey, subject, content } = c.req.valid('json');
-    const prisma = getPrisma(c.env.DATABASE_URL);
-    const note = await prisma.note.create({
-      data: {
-        noteKey,
-        subject,
-        content,
-      },
-    });
-    return note
-      ? c.json(note, 201)
-      : c.json({ code: 404, message: 'Not Found' }, 404);
-  },
-  (result, c) => {
-    if (!result.success) {
-      return c.json({ code: 400, message: 'Validation Error' }, 400);
-    }
-  },
-);
-
-app.doc31('/doc', {
+app.doc('/doc', {
   openapi: '3.1.0',
   info: {
     version: '1.0.0',
     title: 'SprintPost API',
   },
 });
+
+app.route('/', rootApp);
+app.route('/notes', noteApp);
 
 export default app;
