+++ src/utils/prisma.ts
@@ -2,8 +2,7 @@ import { PrismaClient } from '@prisma/client/edge';
 import { withAccelerate } from '@prisma/extension-accelerate';
 
 export const getPrisma = (database_url: string) => {
-  const prisma = new PrismaClient({
+  return new PrismaClient({
     datasourceUrl: database_url,
   }).$extends(withAccelerate());
-  return prisma;
 };
